IT WORKS IT WORK IT WORKS !!!! At least it starts :)

+++ trunk/DofusProtocol/Types/ProtocolTypeManager.cs
@@ -27,7 +27,7 @@ namespace Stump.DofusProtocol.Types
                 if (field != null)
                 {
                     // le cast uint est obligatoire car l'objet n'a pas de type
-                    short id = (short)((uint)field.GetValue(type));
+                    short id = (short)field.GetValue(type);
 
                     m_types.Add(id, type);
